SmartFarm-AI/backend/ai-ml-service/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-multipart>=0.0.6
Pillow>=10.0.0
google-generativeai>=0.8.0
transformers>=4.35.0
torch>=2.0.0
requests>=2.31.0
motor>=3.3.0
pdfplumber>=0.10.3
gTTS>=2.5.0
pydantic>=2.4.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
apscheduler>=3.10.0
